tentativa de criar migrations e models com funcao de ver o carrinho

+++ database/migrations/2021_09_30_181348_create_produtos_table.php
@@ -15,7 +15,7 @@ class CreateProdutosTable extends Migration
     {
         Schema::create('produtos', function (Blueprint $table) {
             $table->id();
-            $table->float('preco', 8, 2)->unsigned();
+            $table->decimal('preco', 8, 2)->unsigned();
             $table->text('descricao');
             $table->integer('qtdestoque')->unsigned();
             $table->string('NomeProduto', 128);
